+++ projects/epgu-constructor/src/lib/component/unique-screen/components/children-clubs/services/program-list/program-list.service.ts
@@ -1,19 +1,9 @@
-import { Injectable } from '@angular/core';
-import { BehaviorSubject, Observable, of } from 'rxjs';
-import {
-  map,
-  shareReplay,
-  switchMap,
-  tap,
-  distinctUntilChanged,
-  catchError,
-  pluck,
-  filter,
-} from 'rxjs/operators';
+import { Injectable, Injector } from '@angular/core';
+import { BehaviorSubject, Observable } from 'rxjs';
+import { map, shareReplay, distinctUntilChanged, pluck } from 'rxjs/operators';
 import { isEqual } from 'lodash';
 import { MicroAppStateQuery } from '@epgu/epgu-constructor-ui-kit';
 import { ListElement } from '@epgu/ui/models/dropdown';
-import { StateService } from '../state/state.service';
 import {
   BaseProgram,
   Filters,
@@ -21,44 +11,16 @@ import {
   ChildrenClubsValue,
   ChildrenClubsState,
   FocusFilter,
-  FindOptionsProgram,
+  VendorType,
 } from '../../models/children-clubs.types';
 import { ScreenService } from '../../../../../../screen/screen.service';
 import { CurrentAnswersService } from '../../../../../../screen/current-answers.service';
 import { ActionService } from '../../../../../../shared/directives/action/action.service';
 import { NEXT_STEP_ACTION } from '../../../../../../shared/constants/actions';
-import { DictionaryApiService } from '../../../../../../shared/services/dictionary/dictionary-api.service';
+import BaseListService from '../base-list.service';
 
 @Injectable()
-export class ProgramListService {
-  public fullLoading$$ = new BehaviorSubject<boolean>(true);
-
-  public fullLoading$ = this.fullLoading$$.asObservable();
-
-  public get fullLoading(): boolean {
-    return this.fullLoading$$.getValue();
-  }
-
-  public loading$$ = new BehaviorSubject<boolean>(false);
-
-  public loading$ = this.loading$$.asObservable();
-
-  public get loading(): boolean {
-    return this.loading$$.getValue();
-  }
-
-  public page$$ = new BehaviorSubject<number>(0);
-
-  public isFinish$$ = new BehaviorSubject<boolean>(false);
-
-  public isFinish$ = this.isFinish$$.asObservable();
-
-  public get isFinish(): boolean {
-    return this.isFinish$$.getValue();
-  }
-
-  public pageSize: number;
-
+export class ProgramListService extends BaseListService<BaseProgram, Filters> {
   public autoScroll$$ = new BehaviorSubject<boolean>(false);
 
   public get autoScroll(): boolean {
@@ -68,12 +30,6 @@ export class ProgramListService {
   public set autoScroll(auto: boolean) {
     this.autoScroll$$.next(auto);
   }
-
-  public data$$ = new BehaviorSubject<BaseProgram[]>([]);
-
-  public data$ = this.data$$.pipe(filter((val) => !!val.length));
-
-  public paginatedData$ = new BehaviorSubject<BaseProgram[]>([]);
 
   public programFilters$ = new BehaviorSubject<Filters>({});
 
@@ -93,63 +49,29 @@ export class ProgramListService {
     shareReplay(1),
   );
 
-  public get data(): BaseProgram[] {
-    return this.data$$.getValue();
-  }
+  protected _args: { nextSchoolYear: boolean; vendor: VendorType; okato: string };
 
-  public load$: Observable<BaseProgram[]> = this.stateService.state$.pipe(
+  protected refetchSubscribtion = this.stateService.state$.pipe(
     distinctUntilChanged(
       (prev, next) =>
         isEqual(prev?.programFilters, next?.programFilters) && prev.okato === next.okato,
     ),
-    map((state) => this.processFilters(state)),
-    tap(() => this.reset()),
-    switchMap((options: FindOptionsProgram) => {
-      options = { ...options, ...this.screenService.component?.arguments };
-      return this.dictionaryApiService
-        .getProgramList({
-          ...options,
-          page: 0,
-          pageSize: 100000,
-        })
-        .pipe(
-          catchError((_) => of([])),
-          tap(() => this.loading$$.next(false)),
-          tap((data: BaseProgram[]) => this.add(data)),
-        );
-    }),
-    shareReplay(1),
+    pluck('programFilters'),
   );
 
+  get disableAutoscroll(): boolean {
+    return this.isFinished.getValue() || this.isLoading.getValue();
+  }
+
   constructor(
-    private dictionaryApiService: DictionaryApiService,
-    private stateService: StateService,
+    protected injector: Injector,
     private appStateQuery: MicroAppStateQuery<ChildrenClubsValue, ChildrenClubsState>,
     private screenService: ScreenService,
     private currentAnswersService: CurrentAnswersService,
     private actionService: ActionService,
   ) {
+    super(injector);
     this.pageSize = (this.screenService.component?.arguments?.pageSize as number) || 10;
-  }
-
-  public isLoaded(): boolean {
-    return false;
-  }
-
-  public getNextPage(): void {
-    const page = this.page$$.getValue() + 1;
-    const data = this.data$$.getValue();
-    const size = page * this.pageSize;
-    const { length } = data;
-    let result: BaseProgram[];
-    if (size >= length) {
-      result = data.slice(size - this.pageSize, length);
-      this.isFinish$$.next(true);
-    } else {
-      result = data.slice(size - this.pageSize, size);
-    }
-    this.paginatedData$.next(this.paginatedData$.getValue().concat(result));
-    this.page$$.next(page);
   }
 
   public selectProgram(uuid): void {
@@ -161,36 +83,16 @@ export class ProgramListService {
     this.actionService.switchAction(NEXT_STEP_ACTION, componentId);
   }
 
-  private add(data: BaseProgram[]): void {
-    if (this.data.length === 0) {
-      this.fullLoading$$.next(false);
-    }
-    if (data.length <= this.pageSize) {
-      this.finish();
-    }
-    this.data$$.next([...this.data].concat(data));
-    this.getNextPage();
+  protected fetchData(filters: Filters): Observable<BaseProgram[]> {
+    return this.apiService.getProgramList({
+      ...this._args,
+      filters,
+      page: 0,
+      pageSize: 100000,
+    });
   }
 
-  private finish(): void {
-    this.isFinish$$.next(true);
-  }
-
-  private resetPagination(): void {
-    this.page$$.next(0);
-    this.paginatedData$.next([]);
-  }
-
-  private reset(): void {
-    this.autoScroll = false;
-    this.isFinish$$.next(false);
-    this.resetPagination();
-    this.fullLoading$$.next(true);
-    this.data$$.next([]);
-  }
-
-  private processFilters(state: ChildrenClubsState): { filters: Filters } {
-    const filters = { ...(state?.programFilters ?? {}) };
+  protected processFilters(filters: Filters = {}): Filters {
     const focus = filters?.focus as ListElement;
     if (focus && focus.id && focus.id !== 'empty-item') {
       filters.focus = focus.id as FocusFilter;
@@ -216,6 +118,6 @@ export class ProgramListService {
       }
     });
     this.programFilters$.next(filters);
-    return { filters };
+    return filters;
   }
 }