+++ projects/epgu-constructor/src/lib/component/unique-screen/components/time-slot/services/smev3/time-slot-smev3.service.ts
@@ -49,11 +49,9 @@ import {
   TimeSlotValueInterface,
   TIMEZONE_STR_OFFSET,
 } from '../../typings';
-
 import { Smev3RestApiService } from '../api/smev3/smev3-rest-api.service';
 import { TimeSlotStateService } from '../state/time-slot-state.service';
 import { TimeSlotSmev3StateService } from '../smev3-state/time-slot-smev3-state.service';
-
 import { TimeSlotErrorService } from '../error/time-slot-error.service';
 import { TimeSlotCalendarService } from '../calendar/time-slot-calendar.service';
 
@@ -68,6 +66,7 @@ export class TimeSlotSmev3Service {
   store$ = combineLatest([
     this.smev3.value$,
     this.smev3.config$,
+    this.smev3.isServiceSpecific$,
     this.smev3.ignoreRootParams$,
     this.smev3.timeSlotRequestAttrs$,
     this.requestListParams$$.pipe(filter((value) => value !== undefined)),
@@ -77,16 +76,20 @@ export class TimeSlotSmev3Service {
     tap(() => this.calendar.isVisibleDays$$.next(true)),
     tap(() => this.state.startLoaded$$.next(false)),
     switchMap(
-      ([data, config, ignoreRootParams, attributes, params]: [
+      ([data, config, isServiceSpecific, ignoreRootParams, attributes, params]: [
         TimeSlotValueInterface,
         TimeSlotsApiItem,
+        boolean,
         string[],
         TimeSlotAttribute[],
         Partial<TimeSlotRequest>,
         null,
       ]) =>
         this.api
-          .getList(this.createRequest(data, config, ignoreRootParams, attributes, params))
+          .getList(
+            this.createRequest(data, config, ignoreRootParams, attributes, params),
+            isServiceSpecific,
+          )
           .pipe(
             catchError(() => {
               this.calendar.isVisibleDays$$.next(false);
@@ -174,16 +177,19 @@ export class TimeSlotSmev3Service {
     concatMap(({ slotList, result }) =>
       iif(
         () => slotList.length > 0,
-        combineLatest([this.smev3.value$, this.smev3.config$]).pipe(
+        combineLatest([this.smev3.value$, this.smev3.config$, this.smev3.isServiceSpecific$]).pipe(
           take(1),
-          concatMap(([data, config]) =>
+          concatMap(([data, config, isServiceSpecific]) =>
             forkJoin(
               slotList.map(({ bookId }) =>
                 this.api
-                  .cancel({
-                    bookId,
-                    eserviceId: (data.eserviceId as string) || config.eserviceId,
-                  })
+                  .cancel(
+                    {
+                      bookId,
+                      eserviceId: (data.eserviceId as string) || config.eserviceId,
+                    },
+                    isServiceSpecific,
+                  )
                   .pipe(catchError((err) => of(err))),
               ),
             ).pipe(
@@ -227,6 +233,7 @@ export class TimeSlotSmev3Service {
         this.smev3.waitingTimeExpired$,
         this.smev3.value$,
         this.smev3.config$,
+        this.smev3.isServiceSpecific$,
         this.smev3.department$,
         this.smev3.attributeNameWithAddress$,
         this.smev3.ignoreRootParams$,
@@ -241,6 +248,7 @@ export class TimeSlotSmev3Service {
             waitingTimeExpired,
             value,
             config,
+            isServiceSpecific,
             department,
             attributeNameWithAddress,
             ignoreRootParams,
@@ -252,6 +260,7 @@ export class TimeSlotSmev3Service {
             boolean,
             TimeSlotValueInterface,
             TimeSlotsApiItem,
+            boolean,
             DepartmentInterface,
             string,
             string[],
@@ -273,6 +282,7 @@ export class TimeSlotSmev3Service {
                   attributes,
                   params,
                 ),
+                isServiceSpecific,
               )
               .pipe(
                 catchError((err) => {