+++ projects/epgu-constructor/src/lib/component/unique-screen/components/time-slots/smev2-time-slots-rest.service.ts
@@ -1,16 +1,14 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { ConfigService } from '@epgu/epgu-constructor-ui-kit';
-
 import { Observable } from 'rxjs';
-
 import { DictionaryOptions } from '@epgu/epgu-constructor-types';
 import { DictionaryResponse } from '../../../../shared/services/dictionary/dictionary-api.types';
 import { ScreenService } from '../../../../screen/screen.service';
 
 @Injectable()
 export class Smev2TimeSlotsRestService {
-  private urlPrefix = this.config.dictionaryUrl;
+  private urlPrefix = this.configService.dictionaryUrl;
 
   private additionalPath =
     this.screenSerivce.component?.attrs?.dictionaryType || 'getAppointment2_mvdr01';
@@ -18,7 +16,7 @@ export class Smev2TimeSlotsRestService {
   constructor(
     private screenSerivce: ScreenService,
     private http: HttpClient,
-    private config: ConfigService,
+    private configService: ConfigService,
   ) {}
 
   public getTimeSlots(requestBody: DictionaryOptions): Observable<DictionaryResponse> {