+++ projects/epgu-constructor/src/lib/component/unique-screen/components/time-slots/smev3-time-slots-rest.service.ts
@@ -10,22 +10,33 @@ import {
   SmevSlotsResponseInterface,
   TimeSlotReq,
 } from './time-slots.types';
+import { ScreenService } from '../../../../screen/screen.service';
 
 @Injectable()
 export class Smev3TimeSlotsRestService {
   private urlPrefix = this.config.mocks.includes('timeSlot')
-    ? `${this.config.mockUrl}/lk/v1/equeue/agg`
+    ? `${this.config.mockUrl}/lk/v1/equeue`
     : this.config.timeSlotApiUrl;
 
-  constructor(private http: HttpClient, private config: ConfigService) {}
+  private isServiceSpecific = this.screenService.component?.attrs?.isServiceSpecific || false;
+
+  constructor(
+    private http: HttpClient,
+    private config: ConfigService,
+    private screenService: ScreenService,
+  ) {}
 
   public getTimeSlots(requestBody: TimeSlotReq): Observable<SmevSlotsResponseInterface> {
-    const path = `${this.urlPrefix}/slots`;
+    const path = `${this.urlPrefix}/${
+      this.isServiceSpecific ? requestBody.eserviceId + '/agg' : 'agg'
+    }/slots`;
     return this.http.post<SmevSlotsResponseInterface>(path, requestBody, { withCredentials: true });
   }
 
   public bookTimeSlot(requestBody: BookTimeSlotReq): Observable<SmevBookResponseInterface> {
-    const path = `${this.urlPrefix}/book?srcSystem=BETA`;
+    const path = `${this.urlPrefix}/${
+      this.isServiceSpecific ? requestBody.eserviceId + '/agg' : 'agg'
+    }/book?srcSystem=BETA`;
     return this.http
       .post<SmevBookResponseInterface>(path, requestBody, { withCredentials: true })
       .pipe(
@@ -33,11 +44,10 @@ export class Smev3TimeSlotsRestService {
       );
   }
 
-  public cancelSlot(requestBody: {
-    eserviceId: string;
-    bookId: string;
-  }): Observable<CancelSlotResponseInterface> {
-    const path = `${this.urlPrefix}/cancel`;
+  public cancelSlot(requestBody: BookTimeSlotReq): Observable<CancelSlotResponseInterface> {
+    const path = `${this.urlPrefix}/${
+      this.isServiceSpecific ? requestBody.eserviceId + '/agg' : 'agg'
+    }/cancel`;
     return this.http.post<CancelSlotResponseInterface>(path, requestBody, {
       withCredentials: true,
     });