+++ projects/epgu-constructor/src/lib/shared/components/upload-and-edit-photo-form/components/photo-error-modal/photo-error-modal.component.ts
@@ -3,8 +3,7 @@ import { ChangeDetectionStrategy, Component, Injector, OnInit } from '@angular/c
 import { ModalBaseComponent, ConfigService } from '@epgu/epgu-constructor-ui-kit';
 import { ImageErrorText } from '../../upload-and-edit-photo-form.model';
 import { imageErrorText } from '../../upload-and-edit-photo-form.constant';
-const imageErrorPicture = require('!raw-loader!projects/epgu-constructor-ui-kit/src/assets/icons/svg/image-error.svg')
-  .default as string;
+
 @Component({
   selector: 'epgu-constructor-photo-error-modal',
   templateUrl: './photo-error-modal.component.html',
@@ -15,7 +14,6 @@ export class PhotoErrorModalComponent extends ModalBaseComponent implements OnIn
   imageErrorText: ImageErrorText;
   customImageErrorText: ImageErrorText;
   imageErrors: string[][];
-  imageErrorPicture = imageErrorPicture;
 
   constructor(public config: ConfigService, public injector: Injector) {
     super(injector);