+++ projects/epgu-constructor/src/lib/shared/components/upload-and-edit-photo-form/components/photo-form-view/photo-form-view.component.ts
@@ -6,16 +6,9 @@ import {
   Input,
   OnInit,
 } from '@angular/core';
-
 import { BaseComponent } from '@epgu/epgu-constructor-ui-kit';
 import { CropTypes, ImageCropOptions } from '@epgu/epgu-constructor-types';
 
-const desktopIcon = require('!raw-loader!projects/epgu-constructor-ui-kit/src/assets/icons/svg/photo-upload-area-desktop.svg')
-  .default as string;
-const mobileIcon = require('!raw-loader!projects/epgu-constructor-ui-kit/src/assets/icons/svg/photo-upload-area-mobile.svg')
-  .default as string;
-const fileIcon = require('!raw-loader!projects/epgu-constructor-ui-kit/src/assets/icons/svg/file-upload-area.svg')
-  .default as string;
 @Component({
   selector: 'epgu-constructor-photo-form-view',
   templateUrl: './photo-form-view.component.html',
@@ -24,6 +17,7 @@ const fileIcon = require('!raw-loader!projects/epgu-constructor-ui-kit/src/asset
 })
 export class PhotoFormViewComponent extends BaseComponent implements OnInit {
   @Input() isDesktop: boolean;
+  @Input() staticDomainAssetsPath: string;
   @Input() croppedImageUrl: string;
   @Input() cropOptions?: ImageCropOptions;
   @Output() hiddenFileInputEvent = new EventEmitter<void>();
@@ -32,19 +26,17 @@ export class PhotoFormViewComponent extends BaseComponent implements OnInit {
   isFacePhotoMode = true;
   uploaderAreaIcon: string;
 
-  public icons = {
-    desktopIcon: desktopIcon,
-    mobileIcon: mobileIcon,
-    fileIcon: fileIcon,
-  };
-
   ngOnInit(): void {
+    const desktopIcon = `${this.staticDomainAssetsPath}/assets/icons/svg/photo-upload-area-desktop.svg`;
+    const mobileIcon = `${this.staticDomainAssetsPath}/assets/icons/svg/photo-upload-area-mobile.svg`;
+    const fileIcon = `${this.staticDomainAssetsPath}/assets/icons/svg/file-upload-area.svg`;
+
     this.isFacePhotoMode = !this.cropOptions || this.cropOptions.cropType === CropTypes.FACE;
 
     this.uploaderAreaIcon = this.isFacePhotoMode
       ? this.isDesktop
-        ? this.icons.desktopIcon
-        : this.icons.mobileIcon
-      : this.icons.fileIcon;
+        ? desktopIcon
+        : mobileIcon
+      : fileIcon;
   }
 }