+++ projects/sf-portal/src/app/app.module.ts
@@ -85,7 +85,3 @@ export function HttpLoaderFactory(httpClient: HttpClient) {
   bootstrap: [AppComponent],
 })
 export class AppModule {}
-
-declare global {
-  type MediaStreamError = Error;
-}